pyproject.toml
[tool.poetry]
name = "trajectopy-core"
version = "1.1.1"
description = "Trajectory Evaluation in Python"
authors = ["Gereon Tombrink <tombrink@igg.uni-bonn.de>"]
maintainers = ["Gereon Tombrink <tombrink@igg.uni-bonn.de>"]
license = "MIT"
keywords = ["trajectory", "evaluation", "alignment", "similarity", "leverarm", "epsg", "robotics"]
readme = "README.md"
homepage = "https://github.com/gereon-t/trajectopy-core"
repository = "https://github.com/gereon-t/trajectopy-core.git"

[tool.poetry.dependencies]
python = ">=3.8.2,<3.12"
autograd = ">=1.6.2"
matplotlib = ">=3.7.3"
networkx = ">=3.1"
pandas = ">=2.0.3"
numpy = ">=1.24.4"
pointset = ">=0.1.5"
PyYAML = ">=6.0.1"
scipy = ">=1.10.1"
rosbags = ">=0.9.16"
rich = ">=13.5.3"

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"
black = "^23.9.1"

[tool.isort]
profile = "black"
line_length = 119

[tool.black]
line-length = 119